// src/app/api/models/digitalTwinModels.ts
export interface Property {
  /**
   * @member {Reported} [reported]
   */
  reported?: Reported;
  /**
   * @member {Desired} [desired]
   */
  desired?: Desired;
}

/**
 * @interface
 * An interface representing InterfaceModel.
 */
export interface InterfaceModel {
  /**
   * @member {string} [name] The name of digital twin interface, e.g.:
   * myThermostat.
   */
  name?: string;
  /**
   * @member {{ [propertyName: string]: Property }} [properties] List of all
   * properties in an interface.
   */
  properties?: { [propertyName: string]: Property };
}

/**
 * @interface
 * An interface representing DigitalTwinInterfaces.
 */
export interface DigitalTwinInterfaces {
  /**
   * @member {{ [propertyName: string]: InterfaceModel }} [interfaces]
   * Interface(s) data on the digital twin.
   */
  interfaces?: { [propertyName: string]: InterfaceModel };
  /**
   * @member {number} [version] Version of digital twin.
   */
  version?: number;
}